# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FITS Backend API
  version: 1.0.0
  description: |
    # FITS (Flexible IT Training System) Backend API

    Production-ready REST API for managing students, teachers, authentication, and signing workflows.

    ## Features
    - JWT-based authentication with refresh tokens
    - Role-based access control (Admin, Teacher, Student)
    - Invitation system for user onboarding
    - Student and teacher management with pagination
    - Digital signing workflow for reports
    - Rate limiting and security headers

    ## Authentication Flow
    1. **Bootstrap Admin**: `POST /api/v1/bootstrap/init` (one-time setup)
    2. **Login**: `POST /api/v1/auth/login` (get access + refresh tokens)
    3. **Use Access Token**: Include `Authorization: Bearer {access_token}` header
    4. **Refresh Token**: `POST /api/v1/auth/refresh` when access token expires
    5. **Logout**: `POST /api/v1/auth/logout` to invalidate tokens

    ## Rate Limiting
    - Global: 100 requests/minute per IP
    - Per-user: Varies by role (Admin: 200/min, Teacher: 100/min, Student: 50/min)

    ## Pagination
    All list endpoints support:
    - `page` (default: 1, min: 1)
    - `limit` (default: 20, max: 100)
  termsOfService: https://fits.example.com/terms
  contact:
    name: API Support Team
    email: support@fits.example.com
    url: https://fits.example.com/support
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/teacher:
    get:
      summary: List teachers with pagination
      description: Retrieve a paginated list of teachers
      operationId: listTeachers
      tags: [Teachers]
      security: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of teachers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTeachersResponse'
              examples:
                success:
                  value:
                    success: true
                    page: 1
                    limit: 20
                    total_count: 45
                    total_pages: 3
                    data:
                      - uuid: 550e8400-e29b-41d4-a716-446655440010
                        email: anna.schmidt@example.com
                        first_name: Anna
                        last_name: Schmidt
                        department: Computer Science
                        created_at: "2025-09-28T10:00:00Z"
                        updated_at: "2025-09-28T10:00:00Z"
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new teacher
      description: |
        Create a new teacher record. Requires admin role.
        Email must be unique. Department is required.
      operationId: createTeacher
      tags: [Teachers]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeacherRequest'
            examples:
              computerScience:
                summary: Computer Science teacher
                value:
                  email: anna.schmidt@example.com
                  first_name: Anna
                  last_name: Schmidt
                  department: Computer Science
              mathematics:
                summary: Mathematics teacher
                value:
                  email: peter.mueller@example.com
                  first_name: Peter
                  last_name: Mueller
                  department: Mathematics
      responses:
        '201':
          description: Teacher created successfully
          headers:
            Location:
              schema:
                type: string
              description: URI of created teacher
              example: /api/v1/teacher/550e8400-e29b-41d4-a716-446655440010
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Teacher'
              examples:
                success:
                  value:
                    success: true
                    message: Teacher created successfully
                    data:
                      uuid: 550e8400-e29b-41d4-a716-446655440010
                      email: anna.schmidt@example.com
                      first_name: Anna
                      last_name: Schmidt
                      department: Computer Science
                      created_at: "2025-10-22T14:30:00Z"
                      updated_at: "2025-10-22T14:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emailConflict:
                  value:
                    success: false
                    code: 409
                    error: Conflict
                    details: A teacher with this email already exists
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/teacher/{uuid}:
    get:
      summary: Get teacher by UUID
      description: Retrieve detailed information about a specific teacher
      operationId: getTeacher
      tags: [Teachers]
      security: []
      parameters:
        - name: uuid
          in: path
          required: true
          description: Teacher UUID
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440010
      responses:
        '200':
          description: Teacher found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Teacher'
              examples:
                success:
                  value:
                    success: true
                    data:
                      uuid: 550e8400-e29b-41d4-a716-446655440010
                      email: anna.schmidt@example.com
                      first_name: Anna
                      last_name: Schmidt
                      department: Computer Science
                      created_at: "2025-09-28T10:00:00Z"
                      updated_at: "2025-09-28T10:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update teacher information
      description: |
        Update an existing teacher's information. Requires admin role.
        Supports partial updates including department changes.
      operationId: updateTeacher
      tags: [Teachers]
      security:
        - BearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          description: Teacher UUID
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440010
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeacherRequest'
            examples:
              updateDepartment:
                summary: Change department
                value:
                  department: Mathematics
              updateAll:
                summary: Update all fields
                value:
                  email: maria.mueller@example.com
                  first_name: Maria
                  last_name: Mueller
                  department: Physics
      responses:
        '200':
          description: Teacher updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Teacher'
              examples:
                success:
                  value:
                    success: true
                    message: Teacher updated successfully
                    data:
                      uuid: 550e8400-e29b-41d4-a716-446655440010
                      email: maria.mueller@example.com
                      first_name: Maria
                      last_name: Mueller
                      department: Physics
                      created_at: "2025-09-28T10:00:00Z"
                      updated_at: "2025-10-22T14:40:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emailConflict:
                  value:
                    success: false
                    code: 409
                    error: Conflict
                    details: Email already in use by another teacher
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete a teacher
      description: |
        Soft delete a teacher from the system. Requires admin role.
        Teacher data is marked as deleted but not physically removed.
      operationId: deleteTeacher
      tags: [Teachers]
      security:
        - BearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          description: Teacher UUID
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440010
      responses:
        '204':
          description: Teacher deleted successfully (no content)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== SIGNING ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication.

        **Format**: `Authorization: Bearer {access_token}`

        **How to get tokens:**
        1. Login: `POST /api/v1/auth/login`
        2. Use returned `access_token` in Authorization header
        3. Refresh when expired: `POST /api/v1/auth/refresh`
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    LimitParam:
      name: limit
      in: query
      required: false
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidJson:
              summary: Invalid JSON
              value:
                success: false
                code: 400
                error: Bad Request
                details: Invalid JSON format
            invalidUuid:
              summary: Invalid UUID format
              value:
                success: false
                code: 400
                error: Bad Request
                details: Invalid UUID format
    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              summary: Missing authorization header
              value:
                success: false
                code: 401
                error: Unauthorized
                details: Missing authorization token
            invalidToken:
              summary: Invalid or expired token
              value:
                success: false
                code: 401
                error: Unauthorized
                details: Invalid or expired token
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            insufficientPermissions:
              summary: Admin role required
              value:
                success: false
                code: 403
                error: Forbidden
                details: Admin role required
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            studentNotFound:
              summary: Student not found
              value:
                success: false
                code: 404
                error: Not Found
                details: Student not found
            teacherNotFound:
              summary: Teacher not found
              value:
                success: false
                code: 404
                error: Not Found
                details: Teacher not found
    ValidationError:
      description: Validation error - invalid field values
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            passwordTooShort:
              summary: Password validation failed
              value:
                success: false
                code: 422
                error: Validation Error
                details: Password must be at least 8 characters long
            invalidEmail:
              summary: Email validation failed
              value:
                success: false
                code: 422
                error: Validation Error
                details: Invalid email format
            missingField:
              summary: Required field missing
              value:
                success: false
                code: 422
                error: Validation Error
                details: "first_name is required"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            serverError:
              summary: Generic server error
              value:
                success: false
                code: 500
                error: Internal Server Error
                details: An unexpected error occurred. Please try again later.
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
        data:
          type: object
          nullable: true
          description: Response payload (varies by endpoint)
    ErrorResponse:
      type: object
      required: [success, code, error]
      properties:
        success:
          type: boolean
          example: false
        code:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type/category
          example: Bad Request
        details:
          type: string
          description: Human-readable error details
          example: Invalid email format
    PaginationMeta:
      type: object
      required: [success, page, limit, total_count, total_pages]
      properties:
        success:
          type: boolean
          example: true
        page:
          type: integer
          description: Current page number
          example: 1
        limit:
          type: integer
          description: Items per page
          example: 20
        total_count:
          type: integer
          description: Total number of items
          example: 150
        total_pages:
          type: integer
          description: Total number of pages
          example: 8

    # ==================== BOOTSTRAP ====================
...
    Teacher:
      type: object
      required: [uuid, email, first_name, last_name, department, created_at, updated_at]
      properties:
        uuid:
          type: string
          format: uuid
          description: Teacher's unique identifier
          example: 550e8400-e29b-41d4-a716-446655440010
        email:
          type: string
          format: email
          description: Teacher's email address
          example: anna.schmidt@example.com
        first_name:
          type: string
          minLength: 1
          maxLength: 100
          description: Teacher's first name
          example: Anna
        last_name:
          type: string
          minLength: 1
          maxLength: 100
          description: Teacher's last name
          example: Schmidt
        department:
          type: string
          minLength: 1
          maxLength: 100
          description: Teacher's department
          example: Computer Science
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2025-09-28T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-09-28T10:00:00Z"
    CreateTeacherRequest:
      type: object
      required: [email, first_name, last_name, department]
      properties:
        email:
          type: string
          format: email
          description: Teacher's email address (must be unique)
          example: anna.schmidt@example.com
        first_name:
          type: string
          minLength: 1
          maxLength: 100
          description: Teacher's first name
          example: Anna
        last_name:
          type: string
          minLength: 1
          maxLength: 100
          description: Teacher's last name
          example: Schmidt
        department:
          type: string
          minLength: 1
          maxLength: 100
          description: Teacher's department
          example: Computer Science
    UpdateTeacherRequest:
      type: object
      description: All fields are optional - only provided fields will be updated
      properties:
        email:
          type: string
          format: email
          description: New email address
          example: maria.mueller@example.com
        first_name:
          type: string
          minLength: 1
          maxLength: 100
          description: New first name
          example: Maria
        last_name:
          type: string
          minLength: 1
          maxLength: 100
          description: New last name
          example: Mueller
        department:
          type: string
          minLength: 1
          maxLength: 100
          description: New department
          example: Physics
    PaginatedTeachersResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationMeta'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Teacher'

    # ==================== SIGNING ====================
...